requirements.txt
aiogram==3.4.1
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
gspread==5.12.4
google-auth==2.27.0
python-dotenv==1.0.0
pydantic==2.6.0
aiohttp==3.9.3
